// src/app/shared/enums.ts
export enum LoadState {
    WAITING = 'indeterminate',
    IDLE = 'determinate'
}

export enum FeedType {
    TOP = 'topstories',
    BEST = 'beststories',
    NEW = 'newstories',
    ASK = 'askstories',
    SHOW = 'showstories',
    JOB = 'jobstories'
}


export enum FeedTypeLabel {
    TOP = 'Top Stories',
    BEST = 'Best Stories',
    NEW = 'New Stories',
    ASK = 'Ask HN',
    SHOW = 'Show HN',
    JOB = 'Jobs HN',
    EMPTY = ''
}

export enum Theme {
    LIGHT = 'light-theme',
    DARK = 'dark-theme'
}

export enum ThemeIcon {
    LIGHT = 'light_mode',
    DARK = 'nights_stay'
}

export enum Palette {
    '#577590' = 0,
    '#43AA8B' = 1,
    '#90BE6D' = 2,
    '#F9C74F' = 3,
    '#F8961E' = 4,
    '#F3722C' = 5,
    '#F94144' = 6
};